Moved loader utility functions to util.js
