+++ src/util.js
@@ -16,6 +16,8 @@
 
 'use strict';
 
+var Module = require('module');
+var fs = require('fs');
 var path = require('path');
 
 /**
@@ -89,7 +91,45 @@ function packageNameFromPath(path) {
   return matches && matches.length > 1 ? matches[1] : null;
 }
 
+/**
+ * Determines the path at which the requested module will be loaded given
+ * the provided parent module.
+ *
+ * @param {string} request The name of the module to be loaded.
+ * @param {object} parent The module into which the requested module will be loaded.
+ */
+function findModulePath(request, parent) {
+  var mainScriptDir = path.dirname(Module._resolveFilename(request, parent));
+  var resolvedModule = Module._resolveLookupPaths(request, parent);
+  var paths = resolvedModule[1];
+  for (var i = 0, PL = paths.length; i < PL; i++) {
+    if (mainScriptDir.indexOf(paths[i]) === 0) {
+      return path.join(paths[i], request.replace('/', path.sep));
+    }
+  }
+  return null;
+}
+
+/**
+ * Determines the version of the module located at `modulePath`.
+ *
+ * @param {?string} modulePath The absolute path to the root directory of the
+ *    module being loaded. This may be null if we are loading an internal module
+ *    such as http.
+ */
+function findModuleVersion(modulePath, load) {
+  if (modulePath) {
+    var pjson = path.join(modulePath, 'package.json');
+    if (fs.existsSync(pjson)) {
+      return load(pjson).version;
+    }
+  }
+  return process.version;
+}
+
 module.exports = {
   stringifyPrefix: stringifyPrefix,
-  packageNameFromPath: packageNameFromPath
+  packageNameFromPath: packageNameFromPath,
+  findModulePath: findModulePath,
+  findModuleVersion: findModuleVersion
 };
